requirements.txt
langchain<=0.2.3
langchain-community<=0.2.3
langchain-openai<=0.1.8
PyPDF2<=3.0.1
python-dotenv<=1.0.1
streamlit<=1.35.0
streamlit-chat<=0.1.1
